FinalProject/elearning_project/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.2
django-filter